requirements.txt
numpy
Pillow
requests
pytesseract
